Changed radiotower influence mechanism to be more fair in early game
Radio towers are now only increasing/decreasing support if the affected side has more than 50% of the town under control. This will stop radio towers from being high important targets in the early game and the money farming in late game.

Cities need to be supplied and missions needs to be done in cities to increase their support during the whole game instead of taking all antennas and instantly win